Feature new sport type

See merge request Android/sport-android!1070

+++ app/src/main/java/org/cxct/sportlottery/network/bet/list/Other.kt
@@ -3,12 +3,11 @@ package org.cxct.sportlottery.network.bet.list
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
-import org.cxct.sportlottery.network.bet.MatchOdd
 
 @JsonClass(generateAdapter = true)
 data class Other (
     @Json(name = "totalAmount")
-    val totalAmount: Int?,
+    val totalAmount: Long?,
     @Json(name = "win")
     val win: Double?,
     @Json(name = "valueBetAmount")
